refactor: logic to hooks
