+++ web/src/components/LogView.tsx
@@ -1,35 +1,14 @@
+import {FC} from 'react';
 import {DeleteIcon} from '@chakra-ui/icons';
-import * as UI from '@components/ui';
-import {useLogBufferContext, useLogContext} from '@middlewares/contexts/useLogContext';
-import {useCallback, useEffect} from 'react';
+import {Alert, AlertIcon, Box, Button, Stack} from '@chakra-ui/react';
+import {OnLogging, useLogView} from './hooks';
 
-export function LogView({
-  onLogging: onLogging,
-}: {
-  onLogging: (num: number) => void;
-}) {
-  const logBuffer = useLogBufferContext();
-  const log = useLogContext();
-  console.log('[LogView]' /*,logBuffer*/);
-
-  useEffect(() => {
-    onLogging(logBuffer.length);
-  }, [logBuffer, onLogging]);
-
-  const clearLog = useCallback(() => {
-    console.log('[LogView] clear');
-    log.clear();
-  }, [log]);
+export const LogView: FC<{onLogging: OnLogging}> = ({onLogging}) => {
+  const {logBuffer, clearLog} = useLogView(onLogging);
 
   return (
-    <UI.Box
-      w="100%"
-      h="100%"
-      p={0}
-      backgroundColor="lightgray"
-      overflowY="hidden"
-    >
-      <UI.Box
+    <Box w="100%" h="100%" p={0} backgroundColor="lightgray" overflowY="hidden">
+      <Box
         height="100%"
         position="absolute"
         right="0"
@@ -38,20 +17,23 @@ export function LogView({
         p={1}
         pr={4}
       >
-        <UI.Button>
+        <Button>
           <DeleteIcon margin="0 auto" onClick={clearLog} />
-        </UI.Button>
-      </UI.Box>
-      <UI.Box h="100%" overflow="scroll">
-        <UI.Stack spacing={3} float="left" width="calc(100% - 64px)" p={2}>
+        </Button>
+      </Box>
+      <Box h="100%" overflow="scroll">
+        <Stack spacing={3} float="left" width="calc(100% - 64px)" p={2}>
           {logBuffer.map((entry) => (
-            <UI.Alert status={entry.type} key={`${entry.timestamp}:${Math.random()}`}>
-              <UI.AlertIcon />[{new Date(entry.timestamp).toISOString()}]:
+            <Alert
+              status={entry.type}
+              key={`${entry.timestamp}:${Math.random()}`}
+            >
+              <AlertIcon />[{new Date(entry.timestamp).toISOString()}]:
               {entry.message}
-            </UI.Alert>
+            </Alert>
           ))}
-        </UI.Stack>
-      </UI.Box>
-    </UI.Box>
+        </Stack>
+      </Box>
+    </Box>
   );
-}
+};
