+++ web/src/components/hooks/index.ts
@@ -3,3 +3,8 @@ export * from './useCommitSessionButton';
 export * from './useFileUploadModal';
 export * from './useGithubAppsButtonGroup';
 export * from './useGithubReposList';
+export * from './useHeaderUserInfo';
+export * from './useLogView';
+export * from './useMarkdownEditor';
+export * from './useMarkdownPreviewer';
+export * from './useMenuBar';
