+++ web/src/components/hooks/useHeaderUserInfo.ts
@@ -1,0 +1,12 @@
+import {useAppContext} from '@middlewares/contexts/useAppContext';
+
+export const useHeaderUserInfo = () => {
+  const app = useAppContext();
+  console.log('user', app.state.user);
+
+  return {
+    user: app.state.user,
+    signIn: app.signIn,
+    signOut: app.signOut,
+  };
+};
